Resolve merge conflicts in sysbench benchmark files:
- Cargo.toml: Keep sysbench_oltp without required-features, add new tpcds/tpcc benchmarks
- sysbench/data.rs: Keep random_ids() and random_range() methods for read-only benchmarks
- sysbench/mod.rs: Update doc comment to include range queries
- sysbench_oltp.rs: Integrate read-only benchmarks with write benchmarks from main

+++ crates/vibesql-executor/tests/conversion_function_tests.rs
@@ -397,9 +397,10 @@ fn test_cast_integer_to_double() {
 
 #[test]
 fn test_cast_double_to_integer() {
-    // MySQL rounds to nearest integer when casting to INTEGER (not truncate)
+    // Default mode is SQLite which truncates toward zero when casting to INTEGER
+    // MySQL mode would round 42.9 to 43, but SQLite truncates to 42
     let result = eval_function("CAST", vec![double_lit(42.9), varchar_lit("INTEGER")]);
-    assert_eq!(result, vibesql_types::SqlValue::Integer(43));
+    assert_eq!(result, vibesql_types::SqlValue::Integer(42));
 }
 
 #[test]
